fix l2_normalize and tests
